requirements.txt
confluent-kafka>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
kafka-python>=2.0.2
pytest>=7.0.0
pytest-mock>=3.10.0
ipaddress>=1.0.23
faker>=18.9.0
termcolor>=2.3.0
tabulate>=0.9.0 